remove getAllFilesRecursively into utility function (had to be in separate file thats only imported from server as can't import 'fs' or 'path' from browser
