+++ lib/mdx.js
@@ -1,12 +1,12 @@
+import MDXComponents from '@/components/MDXComponents'
 import fs from 'fs'
 import matter from 'gray-matter'
-import visit from 'unist-util-visit'
+import renderToString from 'next-mdx-remote/render-to-string'
 import path from 'path'
 import readingTime from 'reading-time'
-import renderToString from 'next-mdx-remote/render-to-string'
-
-import MDXComponents from '@/components/MDXComponents'
+import visit from 'unist-util-visit'
 import imgToJsx from './img-to-jsx'
+import { getAllFilesRecursively } from './utils'
 
 const root = process.cwd()
 
@@ -95,27 +95,6 @@ export async function getFileBySlug(type, slug) {
       ...data,
     },
   }
-}
-
-/**
- *
- * @param  {...string} folders
- * @returns {string[]}
- */
-const getAllFilesRecursively = (...folders) => {
-  const filesAndSubFolders = fs.readdirSync(path.join(...folders))
-
-  const returner = filesAndSubFolders
-    .map((fileOrFolder) => {
-      const fullPath = [...folders, fileOrFolder]
-
-      return fs.statSync(path.join(...fullPath)).isFile()
-        ? path.join(...fullPath)
-        : getAllFilesRecursively(...fullPath)
-    })
-    .reduce((acc, n) => (Array.isArray(n) ? [...acc, ...n] : [...acc, n]), [])
-
-  return returner
 }
 
 export async function getAllFilesFrontMatter(folder) {
