+++ lib/tags.js
@@ -2,7 +2,7 @@ import fs from 'fs'
 import matter from 'gray-matter'
 import path from 'path'
 import { getFiles } from './mdx'
-import { kebabCase } from './utils'
+import { kebabCase } from './utils/kebabCase'
 
 const root = process.cwd()
