+++ lib/utils/files.js
@@ -1,0 +1,28 @@
+import fs from 'fs'
+import path from 'path'
+
+/**
+ *
+ * @param  {...string} folders
+ * @returns {string[]}
+ */
+const getAllFilesRecursively = (...folders) => {
+  // const fs = require('fs')
+  // const path = require('path')
+
+  const filesAndSubFolders = fs.readdirSync(path.join(...folders))
+
+  const returner = filesAndSubFolders
+    .map((fileOrFolder) => {
+      const fullPath = [...folders, fileOrFolder]
+
+      return fs.statSync(path.join(...fullPath)).isFile()
+        ? path.join(...fullPath)
+        : getAllFilesRecursively(...fullPath)
+    })
+    .reduce((acc, n) => (Array.isArray(n) ? [...acc, ...n] : [...acc, n]), [])
+
+  return returner
+}
+
+export { getAllFilesRecursively }
