+++ lib/utils/kebabCase.js
@@ -1,6 +1,8 @@
-export const kebabCase = (str) =>
+const kebabCase = (str) =>
   str &&
   str
     .match(/[A-Z]{2,}(?=[A-Z][a-z]+[0-9]*|\b)|[A-Z]?[a-z]+[0-9]*|[A-Z]|[0-9]+/g)
     .map((x) => x.toLowerCase())
     .join('-')
+
+export { kebabCase as default }
