+++ pages/tags.js
@@ -1,9 +1,9 @@
-import siteMetadata from '@/data/siteMetadata'
-import { kebabCase } from '@/lib/utils'
-import { getAllTags } from '@/lib/tags'
-import Tag from '@/components/Tag'
 import Link from '@/components/Link'
 import { PageSeo } from '@/components/SEO'
+import Tag from '@/components/Tag'
+import siteMetadata from '@/data/siteMetadata'
+import { getAllTags } from '@/lib/tags'
+import { kebabCase } from '@/lib/utils/kebabCase'
 
 export async function getStaticProps() {
   const tags = await getAllTags('blog')
